requirements.txt
pymodbus<3.0.0
pyserial

